[UPDATE] application structure (#15)
