+++ cmd/go-notification/main.go
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2024 TOMIOKA
+//
+// This file is part of the go-notification project.
+
+package main
+
+import "github.com/Nicolas-ggd/go-notification/cmd/go-notification/app"
+
+func main() {
+	app.Run()
+}
