+++ pkg/micro_handlers/micro.go
@@ -1,3 +1,8 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2024 TOMIOKA
+//
+// This file is part of the go-notification project.
+
 package handlers
 
 import (
@@ -42,7 +47,7 @@ func (mh *MicroHandler) BroadcastNotification() micro.HandlerFunc {
 				log.Println(err)
 			}
 
-			// Push the notification into the priority queue
+			// Push the go-notification into the priority queue
 			mh.PriorityQueue.Push(&queue.NotificationHeap{
 				ID:      model.ID,
 				Type:    model.Type,
@@ -72,7 +77,7 @@ func (mh *MicroHandler) ClientBasedNotification() micro.HandlerFunc {
 				log.Println(err)
 			}
 
-			// Push the notification into the priority queue
+			// Push the go-notification into the priority queue
 			mh.PriorityQueue.Push(&queue.NotificationHeap{
 				ID:      model.ID,
 				Type:    model.Type,
@@ -127,7 +132,7 @@ func (mh *MicroHandler) processQueue() {
 			IsView:  notification.IsView,
 		}
 
-		// check if there are specific clients to send the notification to
+		// check if there are specific clients to send the go-notification to
 		if len(notification.Clients) > 0 {
 			// send event to specific clients
 			mh.wss.SendEvent(notification.Clients, model)
